requirements.txt
fastapi
uvicorn[standard]
motor>=3.3
pymongo>=4.6
openai
python-dotenv
pydantic-settings
python-multipart
requests
